--- apps/backoffice/src/app/api/keys/route.ts
@@ -1,11 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-/**
- * @description Retrieve Manage keys
- * TODO:!!!!This is a placeholder!!!!
- * FIXME:!!!!Update the implementation!!!!
- *
- */
-export async function GET(request: NextRequest) {
-  return NextResponse.json({ message: "Not Implemented" });
-}
